resnet-50/requirements.txt
pillow
datasets
transformers
torch
uvicorn[standard]
fastapi
accelerate
sentencepiece